apps/ai_service/requirements.txt
numpy==2.2.6
pillow==11.2.1
pydantic==2.11.4
qdrant-client==1.14.2
redis==5.2.1
requests==2.32.3
python-dotenv==1.0.1
insightface==0.7.3
onnxruntime==1.22.0
